requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
python-multipart
pydantic
jinja2
